+++ rapidfuzz/distance.hpp
@@ -8,6 +8,8 @@
 #include <rapidfuzz/distance/LCSseq.hpp>
 #include <rapidfuzz/distance/Levenshtein.hpp>
 #include <rapidfuzz/distance/OSA.hpp>
+#include <rapidfuzz/distance/Jaro.hpp>
+#include <rapidfuzz/distance/JaroWinkler.hpp>
 
 namespace rapidfuzz {